requirements.txt
Pillow==10.2.0
pycocotools==2.0.7
Requests==2.31.0
torch==2.1.2
torchvision==0.17.1
